fix: rename logger-service to logger.provider
