+++ packages/core/src/provider/index.ts
@@ -1,5 +1,5 @@
 export { IProvider, ProviderManager } from "./provider-manager";
-export { Logger } from "./logger/logger-service";
+export { Logger } from "./logger/logger.provider";
 export { EnvValidatorProvider as Env } from "./environment/env-validator.provider";
 export { ValidateDTO } from "./dto-validator/dto-validator.provider";
 export {
